FIX: code inspection cleanups

+++ maestro-common/src/main/java/org/maestro/common/worker/ThroughputStats.java
@@ -49,7 +49,7 @@ public class ThroughputStats implements PerfStats {
             return 0;
         }
 
-        return count / duration.getSeconds();
+        return (double) count / (double) duration.getSeconds();
     }
 
     @Override
